+++ library/OrderCloud.Catalyst/Integrations/Implementations/Vertex/Models/VertexDiscount.cs
@@ -6,7 +6,7 @@ namespace OrderCloud.Catalyst
 {
 	public class VertexDiscount
 	{
-		public double discountValue { get; set; }
+		public decimal discountValue { get; set; }
 		public VertexDiscountType discountType { get; set; }
 		public string userDefinedDiscountCode { get; set; }
 	}
